% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_CCA.R
\name{CCA}
\alias{CCA}
\title{Canonical Correlation Analysis}
\usage{
CCA(dlist, ...)
}
\arguments{
\item{dlist}{a length-\eqn{N} list of representation matrices. Note that all matrices must have same number of rows that are assumed to be matched observations across different representations.}

\item{...}{optional parameters including\describe{
\item{centering}{a logical to apply centering (default:\code{TRUE}).}
\item{method}{type of summary statistics of the goodness of fit for CCA, either \code{"pillai"} or \code{"yanai"} (default: \code{"yanai"}).}
}}
}
\value{
an \eqn{(N\times N)} matrix of pairwise similarity measures.
}
\description{
The method of Canonical Correlation Analysis (CCA) is one of the most fundamental 
methods to measure association between two matrices. This function implements 
two goodness of fit measures. One is the mean squared CCA correlation, which is 
also known as \emph{Yanai's GCD measure} \insertCite{ramsay_1984_MatrixCorrelation}{repsim}. 
Another is \emph{Pillai's trace}, which is a popular statistic in MANOVA.
}
\examples{
\donttest{
## load the 'twoclass' data
data("twoclass", package="repsim")

# compare Yanai's method (default) with and without centering
Y1 <- CCA(twoclass$data, centering=TRUE)
Y2 <- CCA(twoclass$data, centering=FALSE)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
fields::imagePlot(Y1, xaxt="n", yaxt="n", horizontal=TRUE, main="CCA-centering")
fields::imagePlot(Y2, xaxt="n", yaxt="n", horizontal=TRUE, main="CCA-no centering")
par(opar)
}

}
\references{
\insertAllCited{}
}
\concept{sim}
