% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_DP.R
\name{DP}
\alias{DP}
\title{Dot Product-Based Similarity}
\usage{
DP(dlist, ...)
}
\arguments{
\item{dlist}{a length-\eqn{N} list of representation matrices. Note that all matrices must have same number of rows that are assumed to be matched observations across different representations.}

\item{...}{optional parameters including\describe{
\item{centering}{a logical to apply centering (default:\code{TRUE}).}
}}
}
\value{
an \eqn{(N\times N)} matrix of pairwise similarity measures.
}
\description{
Given two centered representation matrices \eqn{X \in \mathbf{R}^{m\times p}} and \eqn{Y \in \mathbf{R}^{m\times q}}, 
the dot product-based similarity is measured as
\deqn{s(X,Y) = \|X^\top Y \|_F^2,}
which measures cross-variable similarities.
}
\examples{
\donttest{
## load the 'twoclass' data
data("twoclass", package="repsim")

## compute the similarity with and without centering
DP1 <- DP(twoclass$data, centering=TRUE)
DP2 <- DP(twoclass$data, centering=FALSE)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
fields::imagePlot(DP1, xaxt="n", yaxt="n", horizontal=TRUE, main="DP-centering")
fields::imagePlot(DP2, xaxt="n", yaxt="n", horizontal=TRUE, main="DP-no centering")
par(opar)
}

}
\references{
\insertRef{kornblith_2019_SimilarityNeuralNetwork}{repsim}
}
\concept{sim}
