% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_GSL.R
\name{GSL}
\alias{GSL}
\title{Generalized Shape Metric invariant to Linear Transform}
\usage{
GSL(dlist, ...)
}
\arguments{
\item{dlist}{a length-\eqn{N} list of representation matrices. Note that all matrices must have same number of rows that are assumed to be matched observations across different representations.}

\item{...}{optional parameters including\describe{
\item{alpha}{a regularization parameter in \eqn{[0,1]} (default: 0.5).}
\item{centering}{a logical to apply centering (default:\code{FALSE}).}
\item{ndim}{an integer-valued dimension for the common feature space mapping (default: 2).}
}}
}
\value{
an \eqn{(N\times N)} matrix of pairwise similarity measures.
}
\description{
\insertCite{williams_2021_GeneralizedShapeMetrics;textual}{repsim} proposed 
a series of algorithms named as generalized shape metric. This function implements 
the \bold{linear-invariant} metric.
}
\examples{
\donttest{
## load the 'twoclass' data
data("twoclass", package="repsim")

## compare different levels of regularizations
alpha1 <- GSL(twoclass$data, alpha=0)
alpha2 <- GSL(twoclass$data, alpha=0.5)
alpha2 <- GSL(twoclass$data, alpha=1)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3), pty="s")
fields::imagePlot(alpha1, xaxt="n", yaxt="n", horizontal=TRUE, main="GSL-alpha=0")
fields::imagePlot(alpha2, xaxt="n", yaxt="n", horizontal=TRUE, main="GSL-alpha=0.5")
fields::imagePlot(alpha2, xaxt="n", yaxt="n", horizontal=TRUE, main="GSL-alpha=1")
par(opar)
}


}
\references{
\insertAllCited{}
}
\concept{dis}
