% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_GSR.R
\name{GSR}
\alias{GSR}
\title{Generalized Shape Metric invariant to Rotation}
\usage{
GSR(dlist, ...)
}
\arguments{
\item{dlist}{a length-\eqn{N} list of representation matrices. Note that all matrices must have same number of rows that are assumed to be matched observations across different representations.}

\item{...}{optional parameters including\describe{
\item{centering}{a logical to apply centering (default:\code{TRUE}).}
\item{ndim}{an integer-valued dimension for the common feature space mapping (default: 2).}
\item{use.angular}{a logical; \code{FALSE} returns a Procrustes size-and-shape distance with reflections \eqn{d_1} 
and \code{TRUE} to return Kendall's shape space distance \eqn{\theta_1} (default: \code{FALSE}).}
}}
}
\value{
an \eqn{(N\times N)} matrix of pairwise similarity measures.
}
\description{
\insertCite{williams_2021_GeneralizedShapeMetrics;textual}{repsim} proposed 
a series of algorithms named as generalized shape metric. This function implements 
the \bold{rotation-invariant} metric.
}
\examples{
\donttest{
## load the 'twoclass' data
data("twoclass", package="repsim")

## compare Kendall and Procrustes distances
gsrK <- GSR(twoclass$data, use.angular=TRUE)
gsrP <- GSR(twoclass$data, use.angular=FALSE)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
fields::imagePlot(gsrK, xaxt="n", yaxt="n", horizontal=TRUE, main="GSR-Kendall")
fields::imagePlot(gsrP, xaxt="n", yaxt="n", horizontal=TRUE, main="GSR-Procrustes")
par(opar)
}

}
\references{
\insertAllCited{}
}
\concept{dis}
