% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_SVCCA.R
\name{SVCCA}
\alias{SVCCA}
\title{Singular Vector Canonical Correlation Analysis}
\usage{
SVCCA(dlist, ...)
}
\arguments{
\item{dlist}{a length-\eqn{N} list of representation matrices. Note that all matrices must have same number of rows that are assumed to be matched observations across different representations.}

\item{...}{optional parameters including\describe{
\item{centering}{a logical to apply centering (default:\code{TRUE}).}
\item{proportion}{threshold of the amount of variance explained for denoising (default: \eqn{0.95}).}
}}
}
\value{
an \eqn{(N\times N)} matrix of pairwise similarity measures.
}
\description{
Since a standard version of Canonical Correlation Analysis (CCA) often suffers from 
perturbation, denoising based on singular value decomposition is applied as a preprocessing step. 
This is coined as the method of Singular Vector CCA (SVCCA), which measures similarity 
between representation matrices \eqn{[0,1]}. Denoising is applied by the parameter \code{proportion}, 
which truncates the singular vector components by the amount of variance explained.
}
\examples{
\donttest{
## load the 'twoclass' data
data("twoclass", package="repsim")

# compare CCA and SVCCA with two levels of threshold
runCC <- CCA(twoclass$data)
runS1 <- SVCCA(twoclass$data, proportion=0.50)
runS2 <- SVCCA(twoclass$data, proportion=0.95)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3), pty="s")
fields::imagePlot(runCC, xaxt="n", yaxt="n", horizontal=TRUE, main="CCA")
fields::imagePlot(runS1, xaxt="n", yaxt="n", horizontal=TRUE, main="SVCCA thr=50\%")
fields::imagePlot(runS2, xaxt="n", yaxt="n", horizontal=TRUE, main="SVCCA thr=95\%")
par(opar)
}

}
\references{
\insertRef{raghu_2017_SVCCASingularVector}{repsim}
}
\concept{sim}
