% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_twoclass.R
\docType{data}
\name{twoclass}
\alias{twoclass}
\title{Data : Two Class}
\format{
a named list containing\describe{
\item{data}{a list of \eqn{50} landmark representation matrices in \eqn{\mathbf{R}^{100\times 5}}.}
\item{label}{a length-\eqn{50} vector of class labels; either \code{"spiral"} or \code{"cassini"}.}
}
}
\usage{
data(twoclass)
}
\description{
This is a generated data that are perturbed versions of two model shapes, 
\emph{spiral} and \emph{cassini}. The data contains 25 perturbed versions 
of each shape embedded in 5-dimensional space.
}
\examples{
# load the data
data("twoclass", package="repsim")

# take exemplary images from each class
im1 <- twoclass$data[[1]]
im2 <- twoclass$data[[50]]

# apply PCA for low-dimensional embedding
embed1 <- im1\%*\%eigen(stats::cov(im1))$vectors[,1:2]
embed2 <- im2\%*\%eigen(stats::cov(im2))$vectors[,1:2]

# visualize 2-dimensional embeddings
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
plot(embed1, xaxt='n', yaxt='n', xlab='', ylab='', pch=19, cex=0.5, main="spiral")
plot(embed2, xaxt='n', yaxt='n', xlab='', ylab='', pch=19, cex=0.5, main="cassini")
par(opar)

}
\concept{data}
\keyword{datasets}
